Update SM2.0 (NrvO).cps
Updated .cps to give an option to export with a "header" that matches the one output by the Luban Software if used it to create a CNC process. There is data in the Snapmaker header that they use in the "Workspace" where you send to to your machine over wifi. 

I also added the ability to access "Machining Time" and displaying it anyway you'd like with a conversion function. I just thought it would be nice to know how long the machining is going to take.

Finally, I added a way to track "previousFeedRate" because according to the Snapmaker Reference Docs there is no need to write the the federate on every line so instead it only writes it when it changes. <- this implementation does need a second set of eyes incase something was missed and a feed rate isn't being printed correctly which could cause something to happen. This is  a "clean code" implementation not a necessary one so it could be omitted entirely.